CHANGE: Make Catalogue non static and use two AZF/BZF instances
With the static catalogue approach, there is a nasty IndexOutOfBounds Error
during the checking, if a question is correct. The `Catalogue.isCorrect` method
would lookup the question depending on the last catalogue loaded. For example:

The users started the BZF Catalogue (the shorter one) and looks at the
Statistics view, which deserialize a previously completed AZF `Trial`. This
trial would lookup, if each question is correct. And on question 260 of the AZF
Trial the catalogue would throw, because it is currently only initialized with
260 questions.

Making the Catalogue an instance of the Catalogue activity should fix this and
similar future problem.