CHANGE: Remove the State/History only on crashes
If the application crashes, often times this is because the Type of the trials
has changed and connot be parsed correctly anymore. Usually this occurs, when
new questions were added to a catalogue or removed. However we don't want to
delete the users history aggressively on new versions.

Instead a different approach is tried now: Only removing the corrupted states on
a _crash_ of the application.