FIX: Render lines in history correctly corresponding to dates
With the current data structures used to generate the datasets for the history
charts there are a couple of problems:

1. The `List<Trial>` is unordered in respect to the date of each trial. This
can lead to a "jumping" line in the LineChart, where each data point will not
render chronologically on the X axis.

2. The timestamp of the completion of the `Trial` is serialized in PlayerPrefs
with full time information. This leads to generating a separate entry in the
`dates` list regardless if they were actually completed on the same day.

---

The first problem is solved by ordering not only the `List<Date>` but also the
the each list in the `Map<String, List<Trial>>` for each key, based on the
`Trial`'s timestamp.

The second problem is solved by stripping of the time information for any checks
regarding the statistics with the new `createDateOnly()` method. We *do not*
want to save the date in `Trial` in this format, because this might produce
undefined behaviour for existing PlayerPrefs of users.