CHANGE: Refactor Catalogue Layout to ConstraintLayout
With the changes in text size possible for the question and answers it can
easily happen, that the user will choose a large text size which overflows the
display. The questions are very heterogen, meaning there are very short
questions/answer pairs (for which are large size may be suitable) but also very
elaborate and detailed question/answers. For the latter the text will overflow
and the bottom progress bar is not visible anymore.

What we want is to always show the bottom progress bar and the buttons as a
status bar. Only the answers (and maybe the question) should be scrollable.

The simple solution would be to insert a `ScrollView` component. But this
somehow interfers with the relative layout in weird and unpredicable ways. Also
a `LinearLayout` did not help the situation. However it is said anyways, that
the real shit is `ConstraintLayout` so I tried that (Android Studio even helped
with the upgrade process, but this worked horrorbly).

Actually it is quite intuitive and powerful. Now there are contraints which
_always_ render the scroll status bar at the bottom and the answer `RadioGroup`
will adjust its height accordingly and render scrollbars if needed 👌.

See also [here](https://medium.com/quark-works/relative-layout-is-dead-meet-constraint-layout-fa2f8a5fee08)
 for more information