CHANGE: Refactor Activities to Fragments
According to the Material Design Theme for two .. five main views of one's
application it is suggested to use a BottomNavigation bar with fragments in the
main content window.

Therefore all the activities are refactored to fragments _inside_ the Title
activity (which will become the only activity). All the other views get
refactored to become:

* `StatisticsFragment`
* 2x `CatalogueFragment` for BZF and AZF
* `SimulatorFragment`
* `PreferenceFragment`