CHANGE: Add replay button in Toolbar for Simulator & Catalogue
In the new BottomNavigation Layout the _reset_ functionality is missing.
Previoulsy the user was able to reset via the options menu in the app's toolbar
together with some other options.

Now, though, only one option (namely) restart is present for

* CatalogueFragment (BZF)
* CatalogueFragment (AZF)
* Simulator

If the `TitleActivity` shows one of these fragments, it will render a "replay"
icon in the App's toolbar. The individual fragments will handle the reset
accordingly.