CHANGE: Add Tooltip in StatisticFragment for Questions
When the user taps on a question in BZF or AZF (s)he answered already (s)he
probably wants to know, which one this was. For example, if one particular
question was always answered wrong, you want to know which one.

To do this, we generate an Alert Dialog from Material Design and inflate the
same layout as we do for the catalogue (except for the bottom slider).

This required a bit of a refactoring in the layout structure. We now use the
`<inlude layout="@layout/..." />` feature of android.

[Android Documentation](https://developer.android.com/training/improving-layouts/reusing-layouts)