ADD: LabelFormatter to Slider in Catalogue for completed Questions
When scrolling through the catalogue users might want to know, which questions
they already answered and which not. By default, the LabelFormatter of the
Material Slider shows the _Magnitue_ (i.e. the number) of the question. However
this is redundant because the catalogue progress is already shown in a text view
above the slider.

Rather we utilize the LabelFormatter to show a Ballot Box UTF-8 symbol

* ☐ an empty ballot box UTF8 if the question has not been answered yet or
* ☑ a ballot box with check UTF8 if the question has already been answered