CHANGE: { Bottom -> Top } Navigation
Explicit design choice against [Material Design
Guidelines](https://material.io/components/bottom-navigation/).

Personally it looks a bit weird having two user controls at the bottom of the
screen, that is the Slider in CatalogueFragment or the Record button in
SimulatorFragment plus the BottomNavigation. The user is expected to operate the
fragment switching rather infrequent and therefore it is justifiable to place
the navigation component at the top beneath the App's toolbar.