ADD: Vibration Pattern on Edges of answered Questions in Catalogue
When the user scrolls through the catalogue the Tooltip above the Slider will
help her to find unanswered questions. In order to assist this feedback
haptically, we introduce a small _TICK_ vibration on edges of "answeresness".

This means wherever in the catalogue a question changes from answered -> not
answered or vice versa, a small vibration impulse is generated when the user
slides over this part of the catalogue.

Note that we need to do the trick with the extra field `sliderLastQuestion` for
two reasons:

1. This is used to check if the slider actually _changed_. Without this check
the `RadioGroup` will indefinitaly invalide itself at the slider borders.

2. When sliding fastly through the catalogue the value in `onValueChanged()`
will jump drastically, e.g. from 120 .. 140. Instead of only checking the
outlier edges of the new progress we have to check the _range_ between these
two borders for *at least one* edge.