FIX: Preference Dialog Colors for Dark Mode
The problem is, that the dialogs from the EditText- and ListPreferences keep
their default colors (black on white) even with the overwritten dark theme.

Following [this
tutorial](https://medium.com/@JakobUlbrich/building-a-settings-screen-for-android-part-2-2ba63e2d7d1d)
fixes the styles by supplying a custom `layout/edit_text_preference.xml` layout
to the EditTextPreference class, which then uses the correct colors from the
theme.

The buttons in the dialogs were also a bit problematic. For them we set include
custom "AppTheme???.Dialog" and "AppTheme???.Dialog.Button" styles in the root
theme, which customizes the buttons correctly.