ADD: Landscape Layouts for Simulator
The layout of the simulator is a bit clunky in landscape directly because
of the info panel and help texts. So in landscape we use a 30% guide at
which left the info and subtitles are placed and at the right the note
section.

However, since phone layouts have a very short height (in landscape) this
would make it so, that the subtitle help is placed _under_ the record
button. So, we make another version of the layout for phones, which has
the subtitle help panel below the note section.

So everybody should be happy =)