CHANGE: Refactor QuestionFilter to show Percent instead of Absolutes
Abolute numbers in the dialog are easier to understand and to "snap" to. However
there is a problem in the histogram, when a only a subset of the catalogue was
answered. After one such subset trial, some questions were answered N times and
others N+1. However the histogram assumes a uniform scale at the Y-Axis with 1
being equal to the amount of trials conducted.

To me there is no other way to get all trials (partial and complete playlists)
in one histogram by converting the Y-Axis to "Percentage of Correctly answered".