CHANGE: Let Success Rate & Answer Stats be caluclated automatically
Because of random questions the `playlist` does not necessarily match the order
of the `choices` field. Therefore the HashMap + zipping + sorting is needed to
restore the original order of the playlist such that they can be compared with
`choices`

Further Side Effects or this change:

* The `success` field is only used by the "sim" key
* The boolean `corrects` field is not stored in user prefs anymore
* For the SimulationFragment a custom constructor was introduced, which does not
calculates the success rate but takes it as a parameter.