FIX: Initialize all String Resources to prevent Crashes
The annoying "Not attached to Context" Crash seem to happen with the call to
`getString()` from Fragment, while the fragment is not visible anymore. This
seems to be a race condition; some devices have this bug others not.

To prevent this from happening in the future, we initialize all string resources
once while attached to a context, since they don't change over time anyways.